// packages/core/src/compiler/parser/ast.d.ts
export interface AstNodeBase {
    readonly kind: AstNodeKind;
    readonly location: SourceLocation;
}
export declare enum AstNodeKind {
    Program = "Program",
    FunctionDeclaration = "FunctionDeclaration",
    Parameter = "Parameter",
    Block = "Block",
    VariableDeclaration = "VariableDeclaration",
    ReturnStatement = "ReturnStatement",
    IfStatement = "IfStatement",
    ExpressionStatement = "ExpressionStatement",
    AssignmentExpression = "AssignmentExpression",
    BinaryExpression = "BinaryExpression",
    UnaryExpression = "UnaryExpression",
    Identifier = "Identifier",
    NumberLiteral = "NumberLiteral",
    HexLiteral = "HexLiteral",
    BinaryLiteral = "BinaryLiteral",
    BooleanLiteral = "BooleanLiteral",
    ArrayLiteral = "ArrayLiteral",
    ArrayAccess = "ArrayAccess",
    SliceAccess = "SliceAccess",
    TypeAnnotation = "TypeAnnotation"
}
export interface ProgramNode extends AstNodeBase {
    kind: AstNodeKind.Program;
    body: FunctionDeclarationNode[];
}
export interface FunctionDeclarationNode extends AstNodeBase {
    kind: AstNodeKind.FunctionDeclaration;
    name: string;
    parameters: ParameterNode[];
    return_type: TypeAnnotationNode | null;
    body: BlockNode;
    is_exported: boolean;
}
export interface ParameterNode extends AstNodeBase {
    kind: AstNodeKind.Parameter;
    name: string;
    type_annotation: TypeAnnotationNode;
}
export interface BlockNode extends AstNodeBase {
    kind: AstNodeKind.Block;
    statements: StatementNode[];
}
export declare enum TypeName {
    Number = "number",
    Boolean = "boolean",
    NumberArray = "number[]",
    BooleanArray = "boolean[]"
}
export interface TypeAnnotationNode extends AstNodeBase {
    kind: AstNodeKind.TypeAnnotation;
    type_name: TypeName;
    array_size?: number;
}
// ...
export type StatementNode = VariableDeclarationNode | ReturnStatementNode | IfStatementNode | ExpressionStatementNode;